Fix to workaround an OSX 10.6.3 curses library arrow keys bug when running in Interactive (-D) mode. Increased version to 0.9-133.1

+++ squeezeslave/branches/squeezeslave-0.9-133/src/squeezeslave/squeezeslave.c
@@ -601,7 +601,7 @@ int main(int argc, char *argv[]) {
 		      if (FD_ISSET(0, &read_fds)) {
                          while ((key = getch()) != ERR) {
                             ir = getircode(key);
-	                    if (ir == 0x01) {
+	                    if (ir == (unsigned long) 0x01) {
   		               signal_exit_flag = 1;
                             }else{
 			       if (ir != 0) slimproto_ir(&slimproto, 1, 1, ir);
